pre reset
