+++ chains/ethereum/ethereum.go
@@ -11,12 +11,8 @@ func InitializeChain(chainCfg *core.ChainConfig) *core.Chain {
 
 	c := core.NewChain(chainCfg)
 	conn := NewConnection(cfg)
-	c.SetConnection(conn)
-	err := conn.Connect()
-	if err != nil {
-		panic(err)
-	}
 
+	c.SetConnection(conn)
 	emitterContract, err := emitter.NewEmitter(cfg.emitter, conn.conn)
 	if err != nil {
 		panic(err)
