+++ common/helpers.go
@@ -9,6 +9,18 @@ import (
 	"golang.org/x/crypto/ssh/terminal"
 )
 
+const (
+	// In your configuration file, please register as many of the following events
+	DepositAsset = "DepositAsset"
+	NftTransfer  = "NftTransfer"
+	ErcTransfer  = "ErcTransfer"
+
+	// For internal usage only
+	DepositAssetSignature = "DepositAsset(address,bytes32)"
+	NftTransferSignature  = "NFTTransfer(uint256,uint256,address,address,uint256,bytes)"
+	ErcTransferSignature  = "ERCTransfer(uint256,uint256,address,uint256,address)"
+)
+
 // GetPassword displays a message and prompts user to enter a password
 func GetPassword(msg string) []byte {
 	for {
@@ -40,4 +52,18 @@ func FunctionId(sig string) []byte {
 	bytes := []byte(sig)
 	hash := ethcrypto.Keccak256(bytes)
 	return hash[0:4]
+}
+
+func BuildEventSubscriptions(events []string) []string {
+	var arr = []string{}
+	for _, event := range events {
+		if event == DepositAsset {
+			arr = append(arr, DepositAssetSignature)
+		} else if event == NftTransfer {
+			arr = append(arr, NftTransferSignature)
+		} else if event == ErcTransfer {
+			arr = append(arr, ErcTransferSignature)
+		}
+	}
+	return arr
 }
