--- constants/constants.go
@@ -1,27 +1,0 @@
-package constants
-
-const (
-	// In your configuration file, please register as many of the following events
-	DepositAsset = "DepositAsset"
-	NftTransfer  = "NftTransfer"
-	ErcTransfer  = "ErcTransfer"
-
-	// For internal usage only
-	DepositAssetSignature = "DepositAsset(address,bytes32)"
-	NftTransferSignature  = "NFTTransfer(uint256,uint256,address,address,uint256,bytes)"
-	ErcTransferSignature  = "ERCTransfer(uint256,uint256,address,uint256,address)"
-)
-
-func BuildEventSubscriptions(events []string) []string {
-	var arr = []string{}
-	for _, event := range events {
-		if event == DepositAsset {
-			arr = append(arr, DepositAssetSignature)
-		} else if event == NftTransfer {
-			arr = append(arr, NftTransferSignature)
-		} else if event == ErcTransfer {
-			arr = append(arr, ErcTransferSignature)
-		}
-	}
-	return arr
-}
